pyproject.toml
[tool.poetry]
name = "innvestigate"
version = "2.0.2"
description = "A toolbox to innvestigate neural networks' predictions."
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console", 
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
authors = [
    "Maxmilian Alber <workDoTalberDoTmaximilian@gmail.com>",
    "Adrian Hill <hill@tu-berlin.de>",
    "Sebastian Lapuschkin",
    "Miriam Haegele",
    "Kristof Schuett",
    "Philipp Seegerer",
    "Pieter-Jan Kindermans",
    "Henry Webel",
    "Leander Weber",
    "and others",
]
maintainers = ["Adrian Hill <hill@tu-berlin.de>"]
license = "BSD-2-Clause"
readme = "README.md"
homepage = "https://github.com/albermax/innvestigate"
repository = "https://github.com/albermax/innvestigate"
documentation = "https://innvestigate.readthedocs.io/en/latest/"
keywords = ["XAI", "LRP", "Deep Taylor",]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
future = "^0.18.2"
numpy = "^1.22"
matplotlib = "^3.5.1"
scikit-learn = "^1.2.1"
descartes = "^1.1.0"
alphashape = "^1.3.1"
scikit-image = "^0.21.0"
supervision = "^0.16.0"
segment-anything = "^1.0"
torch = "^2.1.1"
torchvision = "^0.16.1"
opencv-python = "^4.8.1.78"
tensorflow-macos = "2.9"
tensorflow-probability = "0.17"
protobuf = "3.20.3"

[tool.poetry.dev-dependencies]
black  = "^22.3"
ruff = "^0.0.212"
Pillow = "^9.0.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
coverage = {extras = ["toml"], version = "^7.0.3"}
codecov = "^2.1.11"
rope = "^1.6.0"
pylint = "^2.12.2"
ipykernel = "^6.19.4"
isort = "^5.10.1"
mypy = "^0.991"
MonkeyType = "^22.2.0"
vulture = "^2.3"
Sphinx = "^6.1.1"
pandas = "^1.3"
ftfy = "^6.1.1"
pyupgrade = "^3.3.1"
pre-commit = "^2.19.0"

[tool.poetry.group.dev.dependencies]
jenkspy = "^0.3.2"
seaborn = "^0.12.2"
jupyter = "^1.0.0"
alphashape = "^1.3.1"
descartes = "^1.1.0"

[tool.isort]
profile = "black"
multi_line_output = 3 
known_first_party = "innvestigate"
known_local_folder = "tests"

[tool.pytest.ini_options]
addopts = "-v -m fast --strict-markers --cov"
# NOTE: locally run tests with `poetry run pytest -m precommit` before a commit 
markers = [
    "precommit",
    "slow",
    "fast",
    "local",
    "graph",
    "application",
    "mnist",
    "reference",
    "layer",
    "ilayers",
    "vgg16",
    "imagenet",
    "resnet50",
    "base",
    "deeptaylor",
    "gradient",
    "init",
    "lrp",
    "wrapper",
    "perturbate",
]

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["innvestigate"]

[tool.coverage.report]
show_missing = false

[tool.ruff]
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "E", # pycodestyle / PEP8
    "F", # pyflakes
    "I", # flake8-isort
    "W", # pycodestyle / PEP8
]
ignore = [
    "E741", # Ignore warning "ambiguous variable name 'l'" as it is useful to iterate over layers
    "B905", # zip's `strict` parameter was only added in Python 3.10
]

[tool.pylint."messages control"]
disable = [
    "raw-checker-failed", 
    "bad-inline-option", 
    "locally-disabled", 
    "file-ignored", 
    "suppressed-message", 
    "useless-suppression", 
    "deprecated-pragma", 
    "use-symbolic-message-instead",
    "no-name-in-module",
    "missing-module-docstring", 
    "missing-class-docstring", 
    "missing-function-docstring", 
    "too-few-public-methods",
]

[tool.pylint.basic]
good-names = ["i", "j", "k", "l", "n", "x", "y", "a", "b", "ex", "Run", "_"]
good-names-rgxs = ["(.*_)?[XYZRABC]s?"]

[tool.mypy]
files = "src/innvestigate"
pretty = true
show_error_codes = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "tensorflow.*",
]
ignore_missing_imports = true

# Build with Poetry
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"